timer5 timer8 asuudaltai busdan tged gaigude tged yu coordinate gehiinmu chigleliin asuudal bga

+++ Core/Inc/swerve_module.h
@@ -37,7 +37,7 @@ typedef struct {
     float integral;
 
     // Steering Motor Parameters
-    uint16_t max_pwm;
+    uint32_t max_pwm;
 
 } SteeringMotor;
